Cast to int64 to circumvent comparison warning.
git-svn-id: http://svn.seqan.de/seqan/trunk@7961 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/swift_local/swift_local.h
@@ -902,7 +902,7 @@ SEQAN_CHECKPOINT
 		TSeed seedOld(seed);
 		extendSeed(seed, host(a), host(b), direction, scoreMatrix, scoreDropOff, GappedXDrop());
 
-		if (getSeedSize(seed) < minLength - (int)floor(minLength*eps))
+		if (static_cast<__int64>(getSeedSize(seed)) < minLength - (int)floor(minLength*eps))
 			return false;
 
 		TPos alignLen = length(row(localAlign, 0)); // TODO not length(row0)!
